Adds maximum Manhattan distance calculation
Implements a function to calculate the maximum Manhattan distance achievable after applying a limited number of changes to a sequence of directions.

The algorithm efficiently determines the maximum distance by tracking the counts of each direction (North, South, East, West) and considering how wasted steps (steps that don't contribute to the Manhattan distance) can be converted into bonus distance with the given number of changes (k).